Refactor aws tags and logs to use common logic
